refaactored commands.py into a commands sub-package to make the code organization cleaner
