+++ bin/workflow
@@ -15,7 +15,7 @@ import traceback
 import argcomplete
 
 from workflow.parser import get_available_tasks, get_available_archives
-from workflow.commands import execute, clean, archive
+from workflow.commands import run, clean, archive
 from workflow.exceptions import CommandLineException, ConfigurationNotFound
 from workflow.notify import notify
 
@@ -40,12 +40,12 @@ subparsers = parser.add_subparsers(
     help='additional help'
 )
 
-######################################################################## execute
-execute_parser = subparsers.add_parser(
+########################################################################### run
+run_parser = subparsers.add_parser(
     'run', 
     help='run the workflow',
 )
-execute_parser.add_argument(
+run_parser.add_argument(
     'task_id',
     metavar='TASK',
     type=str,
@@ -55,12 +55,12 @@ execute_parser.add_argument(
 )
 
 # add a bunch of keyword arguments
-execute_parser.add_argument(
+run_parser.add_argument(
     '-f', '--force',
     action="store_true",
     help="Rerun entire workflow, regardless of task state.",
 )
-execute_parser.add_argument(
+run_parser.add_argument(
     '-d', '--dry-run',
     action="store_true",
     help=(
@@ -68,19 +68,19 @@ execute_parser.add_argument(
         "long it would take."
     ),
 )
-execute_parser.add_argument(
+run_parser.add_argument(
     '-e', '--export',
     action="store_true",
     help="Export shell script instead of running the workflow.",
 )
-execute_parser.add_argument(
+run_parser.add_argument(
     '--notify',
     type=str,
     metavar='EMAIL',
     nargs=1,
     help='Specify an email address to notify on completion',
 )
-execute_parser.set_defaults(func=execute)
+run_parser.set_defaults(func=run)
 
 ########################################################################## clean
 clean_parser = subparsers.add_parser(
@@ -139,6 +139,6 @@ except CommandLineException, e:
     print(e)
     sys.exit(getattr(e, 'exit_code', 1))
 finally:
-    # TODO: should this should be moved inside the execute command?
+    # TODO: should this should be moved inside the run command?
     if getattr(args, 'notify', None):
         notify(*args.notify)
